Add messages endpoints/routes

+++ transport/sanic/endpoints/users/user.py
@@ -1,3 +1,4 @@
+from sanic.exceptions import Unauthorized
 from sanic.request import Request
 from sanic.response import BaseHTTPResponse
 
@@ -11,17 +12,24 @@ from transport.sanic.exceptions import SanicUserNotFound, SanicDBException
 
 
 class UserEndpoint(BaseEndpoint):
-
     async def method_patch(
-            self, request: Request, body: dict, session: DBSession, user_id: int, *args, **kwargs
+            self,
+            request: Request,
+            body: dict,
+            session: DBSession,
+            user_id: int,
+            *args,
+            **kwargs
     ) -> BaseHTTPResponse:
+        if user_id != body["sub"]:
+            raise Unauthorized("user can update only himself")
 
         request_model = RequestPatchUserDto(body)
 
         try:
             db_user = update_user(session, request_model, user_id)
         except DBUserNotExists:
-            raise SanicUserNotFound('User not found')
+            raise SanicUserNotFound("User not found")
 
         try:
             session.commit_session()
@@ -33,13 +41,21 @@ class UserEndpoint(BaseEndpoint):
         return await self.make_response_json(status=200, body=response_model.dump())
 
     async def method_delete(
-            self, request: Request, body: dict, session: DBSession, user_id: int, *args, **kwargs
+            self,
+            request: Request,
+            body: dict,
+            session: DBSession,
+            user_id: int,
+            *args,
+            **kwargs
     ) -> BaseHTTPResponse:
+        if user_id != body["sub"]:
+            raise Unauthorized("user can delete only himself")
 
         try:
             _ = delete_user(session, user_id)
         except DBUserNotExists:
-            raise SanicUserNotFound('User not found')
+            raise SanicUserNotFound("User not found")
 
         try:
             session.commit_session()
@@ -49,17 +65,22 @@ class UserEndpoint(BaseEndpoint):
         return await self.make_response_json(status=204)
 
     async def method_get(
-            self, request: Request, body: dict, session: DBSession, user_id: int, *args, **kwargs
+            self,
+            request: Request,
+            body: dict,
+            session: DBSession,
+            user_id: int,
+            *args,
+            **kwargs
     ) -> BaseHTTPResponse:
+        if user_id != body["sub"]:
+            raise Unauthorized("user can get only himself")
 
         try:
-            _ = get_user_by_id(session, user_id=user_id)
+            db_user = get_user_by_id(session, user_id=user_id)
         except DBUserNotExists:
-            raise SanicUserNotFound('User not found')
+            raise SanicUserNotFound("User not found")
 
-        try:
-            session.commit_session()
-        except (DBDataException, DBIntegrityException) as e:
-            raise SanicDBException(str(e))
+        response_model = ResponseUserDto(db_user)
 
-        return await self.make_response_json(status=204)
+        return await self.make_response_json(body=response_model.dump(), status=200)
